# api.yaml
openapi: 3.0.3
info:
  title: Viking Demo API
  version: 1.0.0
  description: Демонстрационное REST API для работы с таблицей созданных викингов
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
paths:
  /api/vikings:
    get:
      tags:
        - Vikings
      summary: Получить список всех созданных викингов
      operationId: getAllVikings
      responses:
        '200':
          description: Список викингов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Viking'
    post:
      tags:
        - Vikings
      summary: Добавить конкретного викинга
      operationId: addSpecificViking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Viking'
      responses:
        '200':
          description: Викинг успешно добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viking'
  /api/vikings/post:
    post:
      tags:
        - Vikings
      summary: Создать викинга со случайными параметрами
      operationId: createRandomViking
      responses:
        '200':
          description: Викинг успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viking'
  /api/vikings/{id}:
    put:
      tags:
        - Vikings
      summary: Перезаписать параметры конкретного викинга
      operationId: updateViking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Viking'
      responses:
        '200':
          description: Викинг успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viking'
    delete:
      tags:
        - Vikings
      summary: Удалить викинга из таблицы по id
      operationId: deleteViking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Викинг успешно удален
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viking'
  /api/vikings/test:
    get:
      tags:
        - Vikings
      summary: Получить список тестовых викингов
      operationId: getTest
      responses:
        '200':
          description: Список тестовых викингов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
components:
  schemas:
    Viking:
      type: object
      required:
        - name
        - age
        - heightCm
        - hairColor
        - beardStyle
        - equipment
      properties:
        id:
          type: integer
          format: int32
          example: 1
        name:
          type: string
          example: Ragnar
        age:
          type: integer
          format: int32
          example: 34
        heightCm:
          type: integer
          format: int32
          example: 182
        hairColor:
          $ref: '#/components/schemas/HairColor'
        beardStyle:
          $ref: '#/components/schemas/BeardStyle'
        equipment:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentItem'
    EquipmentItem:
      type: object
      required:
        - name
        - quality
      properties:
        name:
          type: string
          example: Axe
        quality:
          type: string
          example: Rare
    HairColor:
      type: string
      enum:
        - Blond
        - Red
        - Brown
        - Black
        - Gray
      example: Blond
    BeardStyle:
      type: string
      enum:
        - SHORT
        - LONG
        - BRAIDED
        - FORKED
        - CLEAN_SHAVEN
      example: BRAIDED